Fix device name display in in-app notices
